Move InferenceResults to server
