+++ x-pack/plugin/ml/src/main/java/org/elasticsearch/xpack/ml/aggs/inference/InternalInferenceAggregation.java
@@ -9,10 +9,10 @@ package org.elasticsearch.xpack.ml.aggs.inference;
 
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
+import org.elasticsearch.inference.InferenceResults;
 import org.elasticsearch.search.aggregations.AggregationReduceContext;
 import org.elasticsearch.search.aggregations.InternalAggregation;
 import org.elasticsearch.xcontent.XContentBuilder;
-import org.elasticsearch.xpack.core.ml.inference.results.InferenceResults;
 
 import java.io.IOException;
 import java.util.List;
@@ -32,7 +32,7 @@ public class InternalInferenceAggregation extends InternalAggregation {
 
     public InternalInferenceAggregation(StreamInput in) throws IOException {
         super(in);
-        inferenceResult = in.readNamedWriteable(InferenceResults.class);
+        inferenceResult = in.readNamedWriteable(org.elasticsearch.inference.InferenceResults.class);
     }
 
     public InferenceResults getInferenceResult() {
